컨트롤러 빈 수정
